requirements.txt
# Core Framework
Flask==3.0.0
Werkzeug==3.0.1

# OCR and Image Processing
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0

# PDF Processing
pdfplumber==0.10.3
PyMuPDF==1.23.8

# Document Export
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4

# Table Detection
tabula-py==2.9.0
camelot-py==0.11.0

# Utilities
numpy==1.26.2